ZNidalee fixes & updates.

+++ ZNidalee/ZNidaleeMenu.h
@@ -12,15 +12,15 @@ class ZNidaleeMenu {
 
         std::map<std::string, IMenuOption *> AllyHealMap;
 
-        std::map<std::string, std::pair<IMenuOption *, int>> HumanQMap;
-        std::map<std::string, std::pair<IMenuOption *, int>> HumanWMap;
-        std::map<std::string, std::pair<IMenuOption *, int>> HumanEMap;
-        std::map<std::string, std::pair<IMenuOption *, int>> HumanRMap;
+        std::map<std::string, std::pair<IMenuOption *, float>> HumanQMap;
+        std::map<std::string, std::pair<IMenuOption *, float>> HumanWMap;
+        std::map<std::string, std::pair<IMenuOption *, float>> HumanEMap;
+        std::map<std::string, std::pair<IMenuOption *, float>> HumanRMap;
 
-        std::map<std::string, std::pair<IMenuOption *, int>> CougarQMap;
-        std::map<std::string, std::pair<IMenuOption *, int>> CougarWMap;
-        std::map<std::string, std::pair<IMenuOption *, int>> CougarEMap;
-        std::map<std::string, std::pair<IMenuOption *, int>> CougarRMap;
+        std::map<std::string, std::pair<IMenuOption *, float>> CougarQMap;
+        std::map<std::string, std::pair<IMenuOption *, float>> CougarWMap;
+        std::map<std::string, std::pair<IMenuOption *, float>> CougarEMap;
+        std::map<std::string, std::pair<IMenuOption *, float>> CougarRMap;
 
         IMenuOption * EnableHealing;
         IMenuOption * ReduceWUsage;
